move: 의존 관계를 고려하여 dto 패키지 이동
