+++ src/main/java/blackjack/dto/CardDto.java
@@ -1,4 +1,4 @@
-package blackjack.domain.dto;
+package blackjack.dto;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardNumber;
