+++ src/main/java/blackjack/dto/DealerResultDto.java
@@ -1,4 +1,4 @@
-package blackjack.domain.dto;
+package blackjack.dto;
 
 import blackjack.domain.gamer.GameResult;
 import java.util.List;
@@ -7,7 +7,7 @@ public record DealerResultDto(int winCount, int loseCount) {
 
     public static DealerResultDto fromPlayerResults(List<GameResult> playerResults) {
         int dealerWinCount = (int) playerResults.stream()
-                .filter(GameResult::isLose)
+                .filter(GameResult::Lose)
                 .count();
 
         int dealerLoseCount = playerResults.size() - dealerWinCount;
