+++ src/main/java/blackjack/dto/PlayerGameResultsDto.java
@@ -1,4 +1,4 @@
-package blackjack.domain.dto;
+package blackjack.dto;
 
 import blackjack.domain.gamer.GameResult;
 import blackjack.domain.gamer.Name;
