+++ src/main/java/blackjack/dto/PlayersHandDto.java
@@ -1,4 +1,4 @@
-package blackjack.domain.dto;
+package blackjack.dto;
 
 import blackjack.domain.gamer.Players;
 import java.util.List;
